+++ examples/strong_get_nekos.rs
@@ -2,7 +2,7 @@
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let resp = nekosbest::st_get::<nekosbest::Neko>().await?;
     let img_url = resp.url();
-    println!("{}", img_url);
+    println!("{img_url}");
     let details = resp.details();
     println!("Artist: {}", details.artist_name);
     println!("Artist link: {}", details.artist_href);