+++ examples/with_amount.rs
@@ -2,6 +2,6 @@
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let resp = nekosbest::get_amount(nekosbest::Category::Neko, 3).await?;
     let img_urls = resp.0;
-    println!("{:?}", img_urls);
+    println!("{img_urls:?}");
     Ok(())
 }